requirements.txt
psutil
openai
anthropic
google-generativeai
requests
python-telegram-bot
mcp
pydantic
python-dotenv
docling
ddgs
